requirements.txt
requests==2.31.0
secure-smtplib==0.1.1
python-dotenv==1.0.0
jinja2==3.1.2